Extract MethodHandlers interface.
